udah bisa nambah akun

+++ database/migrations/2015_05_06_030643_create_donasi.php
@@ -16,7 +16,7 @@ class CreateDonasi extends Migration {
 		{
 			$table->increments('id');
 			$table->timestamp('tanggal')->default(DB::raw('CURRENT_TIMESTAMP'));
-			$table->integer('nominal');
+			$table->bigInteger('nominal');
 			$table->string('termin');
 			$table->string('channel');
 			$table->string('jenis');
